# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "\"赚圈圈API 返回的goods如果其中的goodsin为yes代表它可以进行交易，即：未下架\"",
        "title": "miniproject",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "yyj",
            "email": "2105753640@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/vi",
    "paths": {
...
        "/money/goods/scanning": {
            "get": {
                "description": "\"点击进入商品详情页的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"查询某个商品的详细信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":tables.User,\"user\":tables.User,\"tips\":\"如果goodsin是no则代表已经下架，此时则不显示开启交易按钮\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/homepage": {
            "get": {
                "description": "\"order=1-\u003e返回前十个商品的内容，summary不需要展示出来，是在商品详情页里，主页的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"主页内容\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/message": {
            "get": {
                "description": "\"信息已根据时间按升序排列，id越大越新，消息通知的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "\"返回用户购买、收藏后的信息\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"infor\":[]tables.Message}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error happened"
                    }
                }
            }
        },
...
        "/money/my/cart": {
            "get": {
                "description": "\"购物车\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"查看我的购物车\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"check successfully\",\"infot\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "msg\":\"check successfully\",\"infot\":\"nothing\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/goods": {
            "get": {
                "description": "\"橱窗\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"查看我的橱窗\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"check successfully\",\"infot\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/goods/finish": {
            "get": {
                "description": "\"点击确认完成时的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "\"商家完成订单\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"success\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error happened"
                    }
                }
            }
        },
        "/money/my/goods/unfish": {
            "get": {
                "description": "\"返回订单,需要点完成的是‘my sell’-\u003e[]string 是与我做交易的人的id,因为一个商品可能被多个人购买，所以string切片的长度就是‘完成订单’的订单数,点评价的是‘my buy’\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trade"
                ],
                "summary": "\"返回用户与卖家未完成的订单\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"success\",\"my buy\":[]tables.Good,\"my sell\":[]Two}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/my/message": {
            "get": {
                "description": "\"我的个人信息的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "My"
                ],
                "summary": "\"返回我的信息\"",
                "responses": {
                    "200": {
                        "description": "msg\":\"avatar 是头像对应的url\",\"infor\":tables.User}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\",\"infor\":tables.User}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/search": {
            "post": {
                "description": "\"order=1-\u003e返回前十个商品的内容，summary不需要展示出来-》在商品详情页里,搜索的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"搜索并返回已排序的商品信息\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索框输入的内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"search successfully\",\"infor\":[]tables.Good}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "msg\":\"find nothing\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
...
}